+++ src/operators/add.ml
@@ -38,7 +38,7 @@ let get_again s buf =
   * The [video_init] (resp. [video_loop]) parameter is used to pre-process
   * the first layer (resp. next layers) in the sum; this generalization
   * is used to add either as an overlay or as a tiling. *)
-class add ~kind ~renorm (sources: (int*source) list) video_init video_loop =
+class add ~kind ~renorm (sources: (float*source) list) video_init video_loop =
 object (self)
   inherit operator ~name:"add" kind (List.map snd sources)
 
@@ -92,10 +92,10 @@ object (self)
     (* Compute the list of ready sources, and their total weight *)
     let weight,sources =
       List.fold_left
-        (fun (t,l) (w,s) -> w+t,if s#is_ready then (w,s)::l else l)
-        (0,[]) sources
+        (fun (t,l) (w,s) -> w +. t,if s#is_ready then (w,s)::l else l)
+        (0.,[]) sources
     in
-    let weight = float weight in
+    let weight = weight in
 
     (* Our sources are not allowed to have variable stream kinds.
      * This is necessary, because then we might not be able to sum them
@@ -123,7 +123,7 @@ object (self)
                tmp
              end
            in
-           let c = (float w)/.weight in
+           let c = w/.weight in
 
              if List.length sources = 1 then
                s#get buffer
@@ -199,7 +199,7 @@ let () =
            Only relay metadata from the first source that is effectively \
            summed."
     [ "normalize", Lang.bool_t, Some (Lang.bool true), None ;
-      "weights", Lang.list_t Lang.int_t, Some (Lang.list ~t:Lang.int_t []),
+      "weights", Lang.list_t Lang.float_t, Some (Lang.list ~t:Lang.int_t []),
       Some "Relative weight of the sources in the sum. \
             The empty list stands for the homogeneous distribution." ;
       "", Lang.list_t (Lang.source_t kind_t), None, None ]
@@ -207,11 +207,11 @@ let () =
     (fun p kind ->
        let sources = Lang.to_source_list (List.assoc "" p) in
        let weights =
-         List.map Lang.to_int (Lang.to_list (List.assoc "weights" p))
+         List.map Lang.to_float (Lang.to_list (List.assoc "weights" p))
        in
        let weights =
          if weights = [] then
-           Utils.make_list (List.length sources) 1
+           Utils.make_list (List.length sources) 1.
          else
            weights
        in
@@ -248,7 +248,7 @@ let () =
     ~descr:"Tile sources (same as add but produces tiles of videos)."
     [
       "normalize", Lang.bool_t, Some (Lang.bool true), None ;
-      "weights", Lang.list_t Lang.int_t, Some (Lang.list ~t:Lang.int_t []),
+      "weights", Lang.list_t Lang.float_t, Some (Lang.list ~t:Lang.int_t []),
       Some "Relative weight of the sources in the sum. \
             The empty list stands for the homogeneous distribution." ;
       "proportional", Lang.bool_t, Some (Lang.bool true),
@@ -259,11 +259,11 @@ let () =
     (fun p kind ->
        let sources = Lang.to_source_list (List.assoc "" p) in
        let weights =
-         List.map Lang.to_int (Lang.to_list (List.assoc "weights" p))
+         List.map Lang.to_float (Lang.to_list (List.assoc "weights" p))
        in
        let weights =
          if weights = [] then
-           Utils.make_list (List.length sources) 1
+           Utils.make_list (List.length sources) 1.
          else
            weights
        in